fix #52
